+++ ignite-base/android/app/src/main/java/com/rnbase/MainActivity.java
@@ -1,16 +1,6 @@
 package com.rnbase;
 
-import android.content.Intent;
 import com.facebook.react.ReactActivity;
-import com.airbnb.android.react.maps.MapsPackage;
-import com.i18n.reactnativei18n.ReactNativeI18n;
-import com.oblador.vectoricons.VectorIconsPackage;
-import com.facebook.react.ReactPackage;
-import com.facebook.react.shell.MainReactPackage;
-import com.learnium.RNDeviceInfo.RNDeviceInfo;
-
-import java.util.Arrays;
-import java.util.List;
 
 public class MainActivity extends ReactActivity {
 
@@ -22,36 +12,4 @@ public class MainActivity extends ReactActivity {
     protected String getMainComponentName() {
         return "RNBase";
     }
-
-    /**
-     * Returns whether dev mode should be enabled.
-     * This enables e.g. the dev menu.
-     */
-    @Override
-    protected boolean getUseDeveloperSupport() {
-        return BuildConfig.DEBUG;
-    }
-
-    /**
-     * A list of packages used by the app. If the app uses additional views
-     * or modules besides the default ones, add more packages here.
-     */
-    @Override
-    protected List<ReactPackage> getPackages() {
-        return Arrays.<ReactPackage>asList(
-            new MainReactPackage(),
-            new MapsPackage(this),
-            new RNDeviceInfo(),
-            new ReactNativeI18n(),
-            new VectorIconsPackage()
-        );
-    }
-
-    // Add onNewIntent
-    @Override
-    public void onNewIntent (Intent intent) {
-      super.onNewIntent(intent);
-
-    }
-
 }
