+++ ignite-base/android/app/src/main/java/com/rnbase/MainApplication.java
@@ -1,0 +1,46 @@
+package com.rnbase;
+
+import android.app.Application;
+import android.util.Log;
+
+import com.facebook.react.ReactApplication;
+import com.facebook.react.ReactInstanceManager;
+import com.facebook.react.ReactNativeHost;
+import com.facebook.react.ReactPackage;
+import com.facebook.react.shell.MainReactPackage;
+
+// 3rd party imports
+import com.learnium.RNDeviceInfo.RNDeviceInfo;
+import com.airbnb.android.react.maps.MapsPackage;
+import com.i18n.reactnativei18n.ReactNativeI18n;
+import com.oblador.vectoricons.VectorIconsPackage;
+
+import java.util.Arrays;
+import java.util.List;
+
+public class MainApplication extends Application implements ReactApplication {
+
+  private final ReactNativeHost mReactNativeHost = new ReactNativeHost(this) {
+    @Override
+    protected boolean getUseDeveloperSupport() {
+      return BuildConfig.DEBUG;
+    }
+
+    @Override
+    protected List<ReactPackage> getPackages() {
+
+      return Arrays.<ReactPackage>asList(
+          new MainReactPackage(),
+          new MapsPackage(),
+          new RNDeviceInfo(),
+          new ReactNativeI18n(),
+          new VectorIconsPackage()
+      );
+    }
+  };
+
+  @Override
+  public ReactNativeHost getReactNativeHost() {
+      return mReactNativeHost;
+  }
+}
